UF-2441 [FIX] Procurement batch cron: Sometimes the rules are launched two times

+++ procurement_report/procurement_batch.py
@@ -128,16 +128,25 @@ class procurement_batch_cron(osv.osv):
         if isinstance(ids, (int, long)):
             ids = [ids]
         
-        for batch in self.browse(cr, uid, ids, context=context):
-            if vals.get('type') and vals.get('type') != batch.type:
-                for cron in batch.cron_ids:
-                    cron_obj.unlink(cr, uid, cron.id, context=context)
-                self._create_associated_cron(cr, uid, batch.id, vals, context=context)
-            else:
-                for cron in batch.cron_ids:
-                    if 'name' in vals:
-                        vals.pop('name')
-                    cron_obj.write(cr, uid, cron.id, vals, context=context)
+        if vals is None:
+            vals = {}
+
+        cron_vals = vals.copy()
+        if cron_vals.get('last_run_on'):
+            del cron_vals['last_run_on']
+
+        if cron_vals:
+            for batch in self.browse(cr, uid, ids, context=context):
+                if vals.get('type') and vals.get('type') != batch.type:
+                    for cron in batch.cron_ids:
+                        cron_obj.unlink(cr, uid, cron.id, context=context)
+                    self._create_associated_cron(cr, uid, batch.id, vals, context=context)
+                else:
+                    for cron in batch.cron_ids:
+                         if 'name' in vals:
+                            vals.pop('name')
+                         print 'update_cron %s' % cron.function
+                         cron_obj.write(cr, uid, cron.id, vals, context=context)
                     
         return super(procurement_batch_cron, self).write(cr, uid, ids, vals, context=context)
